Lots of refactoring...
