+++ __init__.py
@@ -1,3 +1,6 @@
-from ball_classifier import BallClassifier
-from ball_labeller import BallLabeller
-from classifier_metrics import ClassifierMetrics
+from BallClassifier import BallClassifier
+from BallLabeller import BallLabeller
+from ClassifierMetrics import ClassifierMetrics
+from TrajectoryPredictor import TrajectoryPredictor
+from VideoInterface import VideoInterface
+from WebcamStream import WebcamStream
