+++ range-detector.py
@@ -19,7 +19,7 @@ def callback(value):
 def setup_trackbars(range_filter):
     cv2.namedWindow("Trackbars", 0)
 
-    for i in ["MIN", "MAX"]:    
+    for i in ["MIN", "MAX"]:
         v = 0 if i == "MIN" else 255
 
         for j in range_filter:
@@ -88,9 +88,11 @@ def main():
             else:
                 frame_to_thresh = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
 
-        v1_min, v2_min, v3_min, v1_max, v2_max, v3_max = get_trackbar_values(range_filter)
+        v1_min, v2_min, v3_min, v1_max, v2_max, v3_max = get_trackbar_values(
+            range_filter)
 
-        thresh = cv2.inRange(frame_to_thresh, (v1_min, v2_min, v3_min), (v1_max, v2_max, v3_max))
+        thresh = cv2.inRange(
+            frame_to_thresh, (v1_min, v2_min, v3_min), (v1_max, v2_max, v3_max))
 
         if args['preview']:
             preview = cv2.bitwise_and(image, image, mask=thresh)
